test 2 db

+++ db/copy_database.sh
@@ -5,13 +5,15 @@ CONTAINER_NAME="postgres"
 DB_USER="licentauser"
 DB_NAME="licentadb"
 BACKUP_PATH="/backup.sql"
-HOST_BACKUP_PATH="/home/david/Desktop/LICENTA/licenta-be/db/backup_host.sql"
+HOST_BACKUP_PATH="/home/david/david/LICENTA/BE/db/backup_host.sql"
 
 # Dump the database inside the container
-docker exec $CONTAINER_NAME pg_dump -U $DB_USER -d $DB_NAME -f $BACKUP_PATH
+if docker exec $CONTAINER_NAME pg_dump -U $DB_USER -d $DB_NAME --clean -f $BACKUP_PATH; then
+   echo "Dump successful, copying to host..."
 
-# Copy the dump file to the host machine
-docker cp $CONTAINER_NAME:$BACKUP_PATH $HOST_BACKUP_PATH
-
-# Print completion message
-echo "Database dump completed and copied to $HOST_BACKUP_PATH"
+  # Copy the dump file to the host machine
+  docker cp $CONTAINER_NAME:$BACKUP_PATH $HOST_BACKUP_PATH
+  echo "Database dump completed and copied to $HOST_BACKUP_PATH"
+else
+  echo "failed to dump database"
+fi
